requirements.txt
grpcio==1.51.1
grpcio-tools==1.51.1
hazelcast-python-client==5.1
protobuf==4.21.12
